do some renaming of components and simplify auth code
