+++ lib/src/game_schedule/game_card_component.css
