+++ lib/src/game_schedule/game_schedule_component.dart
@@ -5,20 +5,21 @@ import 'package:angular_components/content/deferred_content.dart';
 import 'package:angular_components/material_button/material_button.dart';
 import 'package:angular_components/material_icon/material_icon.dart';
 import 'package:angular_components/material_toggle/material_toggle.dart';
+import 'package:firebase/firebase.dart' as fb;
 import 'package:angular_router/angular_router.dart';
 import 'package:nhl/src/router/route_paths.dart';
-import 'package:nhl/src/today/game_card_component.dart';
-import 'package:nhl/src/today/today_list_service.dart';
+import 'package:nhl/src/game_schedule/game_card_component.dart';
+import 'package:nhl/src/game_schedule/game_schedule_service.dart';
 import 'package:nhl/src/models/game.dart';
 import 'package:nhl/src/router/routes.dart';
 
 @Component(
-  selector: 'today-list',
+  selector: 'game_schedule-list',
   styleUrls: [
-    'today_list_component.css',
+    'game_schedule_component.css',
     'package:angular_components/app_layout/layout.scss.css',
   ],
-  templateUrl: 'today_list_component.html',
+  templateUrl: 'game_schedule_component.html',
   directives: [
     coreDirectives,
     DeferredContentDirective,
@@ -53,9 +54,17 @@ class TodayListComponent implements OnActivate {
   Future<NavigationResult> goToTomorrow() =>
       _router.navigate(tomorrowUrl);
 
+  bool isAuthenticated() => fb.auth().currentUser != null;
   // This isn't particularly nice or elegant but I can't figure out how to do
   // it nicely and I'm admitting defeat to AngularRouter
   void onActivate(_, RouterState current) async {
+
+    // check if we are signed in - if not, redirect to login
+    if (!isAuthenticated()) {
+      _router.navigate('/login');
+      return;
+    }
+
     final String date = getDateFromMap(current.parameters);
 
     today = DateTime
